Created the view for form_fragment
