+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.PMPDPlayGround">
         <activity
-            android:name="com.tijanidian.ex06form.Ut02ex06Activity"
+            android:name="com.tijanidian.ex06form.presentation.Ut02ex06Activity"
             android:exported="true" />
         <activity
             android:name="com.tijanidian.mvvm.presentation.MvvmActivity"
