--- app/src/main/java/com/tijanidian/ex06form/FormFragment.kt
@@ -1,23 +1,0 @@
-package com.tijanidian.ex06form
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-
-class FormFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
-    companion object {
-        fun createInstance() = FormFragment
-    }
-}
