--- app/src/main/java/com/tijanidian/ex06form/ListFragment.kt
@@ -1,27 +1,0 @@
-package com.tijanidian.ex06form
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-
-class ListFragment:Fragment() {
-
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
-    companion object{
-        val TAG:String=Ut02ex06Activity::class.java.simpleName
-        fun createInstance()=ListFragment()
-
-    }
-
-}
