+++ app/src/main/res/menu/main_manu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/action_add"
+        android:icon="@drawable/ic_leon"
+        android:title="Add"
+        app:showAsAction="ifRoom" />
+   
+</menu>
