+++ app/src/main/res/values/colors.xml
@@ -1,13 +1,33 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+
+    <color name="greey">#222222</color>
+    <color name="blue">#1784F3</color>
+
+
+    <!-- Styles Colors Theme -->
+    <color name="primaryColor">#304ffe</color>
+    <color name="primaryLightColor">#7a7cff</color>
+    <color name="primaryDarkColor">#0026ca</color>
+    <color name="primaryTextColor">#ffffff</color>
+    <color name="secondaryColor">#00c853</color>
+    <color name="secondaryLightColor">#5efc82</color>
+    <color name="secondaryDarkColor">#009624</color>
+    <color name="secondaryTextColor">#000000</color>
+
+    <!-- Hex. Colors -->
+    <color name="green">#1b5e20</color>
+    <color name="light_green">#4c8c4a</color>
+    <color name="dark_green">#003300</color>
+    <color name="black">#FF000000</color>
+    <color name="white">#FFFFFFFF</color>
+    <color name="light_grey">#E3E3E3</color>
+    <color name="light_orange">#ffc046</color>
+    <color name="dark_orange">#c56000</color>
+    <color name="orange">#ff8f00</color>
     <color name="purple_200">#FFBB86FC</color>
     <color name="purple_500">#FF6200EE</color>
     <color name="purple_700">#FF3700B3</color>
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
-    <color name="greey">#222222</color>
-    <color name="blue">#1784F3</color>
-
 </resources>
