+++ app/src/main/res/values/communities_array.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="communities_array">
+        <item>Castilla y León</item>
+        <item>Castilla y la Mancha</item>
+        <item>Andalucia</item>
+        <item>Aragón</item>
+        <item>Asturias</item>
+        <item>Baleares</item>
+        <item>Cantabria</item>
+        <item>Comunidad Valenciana</item>
+        <item>Extremadura</item>
+        <item>Galicia</item>
+        <item>La Rioja</item>
+        <item>Madrid</item>
+        <item>Murcia</item>
+        <item>Navarr</item>
+        <item>País Vasco</item>
+    </string-array>
+</resources>
