+++ app/src/main/res/values/strings.xml
@@ -30,4 +30,7 @@
     <string name="glovo_secundario_text">Las más famosas</string>
     <string name="label_price">2,50€</string>
     <string name="text_right_icon">20-25 min</string>
+
+
+
 </resources>
